pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thistlebot"
version = "0.1.0"
description = "Lightweight local AI assistant with a gateway and CLI"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Ibrahim El Nemr" }]
dependencies = [
  "typer>=0.12",
  "httpx>=0.27",
  "fastapi>=0.110",
  "uvicorn>=0.30",
  "pyyaml>=6.0",
]

[project.scripts]
thistlebot = "thistlebot.cli:app"

[tool.setuptools]
package-dir = { "" = "." }

[tool.setuptools.packages.find]
where = ["."]
include = ["thistlebot*"]

[tool.setuptools.package-data]
thistlebot = ["config/*.yaml"]
